+++ com.unity.shadergraph/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Next version
 
+### HD Render Pipeline support
+
+![](.data/hd_render_pipeline.png)
+
+Shader Graph now supports the High Definition Render Pipeline with both PBR and Unlit Master nodes. Shaders built with Shader Graph work with both the Lightweight and HD render pipelines.
+
+### Vertex position
+
+![](.data/vertex_position.png)
+
+You can now modify vertex position via the Position slot on the PBR and Unlit Master nodes. By default, the input to this node is object space position. Custom inputs to this slot should specify the absolute local position of a given vertex. Certain nodes (such as Procedural Shapes) are not viable in the vertex shader. Such nodes are incompatible with this slot.
+
 ### Master node settings
 
 ![](.data/menu_settings.png)
@@ -21,6 +33,12 @@ In the expanded property window, you can now also toggle if the property is expo
 
 You can now change the path of Shader Graphs and Sub Graphs. When you change the path of a Shader Graph, this modifies the location it has in the shader selection list. When you change the path of Sub Graph, it will have a different location in the node creation menu.
 
+### Is Front Face node
+
+![](.data/face_sign.png)
+
+With this node, you can change graph output depending on the face sign of a given fragment. If the current fragment is part of a front face, the node returns True. For a back face, the node returns False.
+Note: This functionality requires that you have enabled **two sided** on the Master node.
 
 ### Gradient nodes
 
@@ -33,6 +51,18 @@ This adds gradient functionality via two new nodes. The Sample Gradient node sam
 ![](.data/sphereMask.png)
 
 With this node, you can create a sphere mask. Indicate the starting coordinate and center point, and the sphere mask will use these with the **Radius** and **Hardness** parameters. Sphere mask functionality works in both 2D and 3D spaces, and is based on the vector coordinates in the **Coords and Center** input.
+
+### Texture3D and Texture2D Array
+
+![](.data/texture_nodes.png)
+
+This change expands Unity's support for Texture types via two new property types and four new nodes. These allow you to define and sample Texture 3D and Texture 2D Array type assets in Shader Graph.
+
+### Texture 2D LOD node
+
+![](.data/texture_2d_lod_node.png)
+
+This adds a new node for LOD functionality on a Texture 2D Sample. Sample Texture 2D LOD uses the exact same input and output slots as Sample Texture 2D, but also includes an input for level of detail adjustments via a Vector1 slot. 
 
 ### Show generated code
 
@@ -52,3 +82,4 @@ You can now see the generated code for any specific node. To do so, right-click 
 - Your system locale can no longer cause incorrect commands due to full stops being converted to commas.
 - Deserialization of subgraphs now works correctly.
 - Sub graphs are now suffixed with (sub), so you can tell them apart from other nodes.
+- The preview of a node does not obstruct the selection outliner anymore.